monoprior: stereo_catalog streams to the Rerun viewer instead of registering a layer (PromptDA Polycam pattern)
Replaces stereo_catalog_layer: RerunTyroConfig output (spawn/save/connect), catalog video_time timeline, raw front
videos relayed with send_columns, per-frame world_T_rig on the rig node, rectified image + EncodedDepthImage, TSDF
mesh re-logged every --mesh-every frames so scrubbing shows it grow, final mesh at the end. Task
monoprior-stereo-catalog.
