+++ packages/monoprior/tools/apps/stereo_catalog.py
@@ -1,0 +1,6 @@
+import tyro
+
+from monopriors.apis.stereo_catalog import StereoCatalogConfig, main
+
+if __name__ == "__main__":
+    main(tyro.cli(StereoCatalogConfig))
