--- packages/monoprior/tools/apps/stereo_catalog_layer.py
@@ -1,6 +1,0 @@
-import tyro
-
-from monopriors.apis.stereo_catalog_layer import StereoCatalogLayerConfig, main
-
-if __name__ == "__main__":
-    main(tyro.cli(StereoCatalogLayerConfig))
