+++ pixi.toml
@@ -445,11 +445,11 @@ depends-on = ["_monoprior-download-stereo"]
 cwd = "packages/monoprior"
 description = "Launch the stereo depth Gradio app (LiteAnyStereo V2 + Rerun rig viewer)"
 
-[feature.monoprior.tasks.monoprior-stereo-catalog-layer]
-cmd = "python tools/apps/stereo_catalog_layer.py"
+[feature.monoprior.tasks.monoprior-stereo-catalog]
+cmd = "python tools/apps/stereo_catalog.py"
 cwd = "packages/monoprior"
 env = { RERUN_INSECURE_SKIP_HOST_CHECK = "1" }
-description = "Rectify a catalog rig's front fisheye pair, run LiteAnyStereo V2, register the depth as a layer (robocap on :51235 by default)"
+description = "Stream LiteAnyStereo V2 depth + an incremental TSDF mesh for a catalog rig segment (robocap front pair on :51235) into the Rerun viewer"
 
 [feature.monoprior.tasks.monoprior-relative-depth]
 cmd = "python tools/demos/relative_depth.py --image-path data/examples/single-image/room.jpg"
